fix: commission rate selection

+++ database/migrations/2021_10_13_130318_create_overall_commission_rates_table.php
@@ -15,7 +15,7 @@ class CreateOverallCommissionRatesTable extends Migration
     {
         Schema::create('overall_commission_rates', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('rate')->nullable();
+            $table->float('rate')->nullable();
             $table->timestamps();
         });
     }
